Refactored old integration tests to use integration testing helpers. Extended integration testing helpers to allow for configurable test contexts.
