+++ packages/test/package.json
@@ -8,7 +8,9 @@
     "build:ts": "tsc --build",
     "build:protos": "node ./scripts/compile-proto.js",
     "test": "ava ./lib/test-*.js",
-    "test.watch": "ava --watch ./lib/test-*.js"
+    "test.watch": "ava --watch ./lib/test-*.js",
+    "test-old": "RUN_INTEGRATION_TESTS=1 ava ./lib/test-integration-old.js",
+    "test-refactor": "RUN_INTEGRATION_TESTS=1 ava ./lib/test-integration-new.js"
   },
   "ava": {
     "timeout": "60s",
