+++ packages/test/src/workflows/success-string.ts
@@ -1,3 +1,4 @@
+// TODO(thomas): worth removing?
 export async function successString(): Promise<string> {
   return 'success';
 }
